Keep converting commands to use CommandHandler
